Tidied up codebase
